pyproject.toml
[build-system]
requires = ["setuptools >= 61.0", "wheel", "Cython"]
build-backend = "setuptools.build_meta"

[project]
name = "lego-lcd"
version = "0.1"
readme = "README.md"
requires-python = ">= 3.8"
dependencies = ["python-networkmanager"]

[project.scripts]
lcd-clock = "lego_lcd.clock:main"
